+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/citizen/Login/index.js
@@ -499,6 +499,7 @@ const Login = ({ stateCode, isUserRegistered = true }) => {
               validationConfig={validationConfig}
               allValidationConfigs={allValidationConfigs}
               enableUserPreferences={enableUserPreferences}
+              countryCode={params.countryCode || ""}
             />
           </Route>
           <Route path={`${path}/otp`}>